remove try catch block

+++ controller/deleteUser.js
@@ -1,13 +1,10 @@
 const userModel = require('../models/userModel');
 async function deleteUser(req, res) {
     const {id} = req.query;
-    try {
-        await userModel.deleteOne({_id: id});
-        res.status(200).json({ID: id, Message: "User deleted successfully!"});
-    } catch (e) {
-        console.log('Not deleted User!');
-        res.json({status: "Unsuccessful!", Message: e.message});
-    }
+    const user = await userModel.findOne({_id:id});
+    if(!user) throw "User not found";
+    await userModel.deleteOne({_id: id});
+    res.status(200).json({ID: id, Message: "User deleted successfully!"});
 }
 
 module.exports = deleteUser;
